+++ app/src/main/java/org/simple/clinic/contactpatient/ContactPatientBottomSheet.kt
@@ -336,6 +336,14 @@ class ContactPatientBottomSheet : BaseBottomSheet<
     callPatientView.showPatientWithNoPhoneNumberResults = true
   }
 
+  override fun showDeadPatientStatus() {
+    callPatientView.showPatientDiedStatus = true
+  }
+
+  override fun hideDeadPatientStatus() {
+    callPatientView.showPatientDiedStatus = false
+  }
+
   private fun backPressed() {
     hotEvents.onNext(BackClicked)
   }