+++ app/src/main/java/org/simple/clinic/contactpatient/ContactPatientUi.kt
@@ -35,6 +35,9 @@ interface ContactPatientUi {
 
   fun showPatientWithNoPhoneNumberResults()
 
+  fun showDeadPatientStatus()
+  fun hideDeadPatientStatus()
+
   /**
    * Select reminder view
    */