+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Add diabetes treatment question in medical history screen
 - Hide resend sms button when OTP attempts are blocked
 - Show patient died status in patient summary screen
+- Show patient died status in contact patient bottom sheet
 - [In Progress: 19 Aug 2021] UI improvements for medical history screen
   - Show hypertension diagnosis and treatment in single card
   - Show separate cards for hypertension and diabetes diagnosis